move direct components of TorService to their own components package
