+++ topl-service/src/main/java/io/matthewnelson/topl_service/service/components/ActionCommands.kt
@@ -64,7 +64,7 @@
 *     modified version of TorOnionProxyLibrary-Android, and you must remove this
 *     exception when you distribute your modified version.
 * */
-package io.matthewnelson.topl_service.service
+package io.matthewnelson.topl_service.service.components
 
 import android.content.Intent
 import io.matthewnelson.topl_service.TorServiceController
