+++ topl-service/src/main/java/io/matthewnelson/topl_service/service/components/ServiceActionProcessor.kt
@@ -64,11 +64,12 @@
 *     modified version of TorOnionProxyLibrary-Android, and you must remove this
 *     exception when you distribute your modified version.
 * */
-package io.matthewnelson.topl_service.service
+package io.matthewnelson.topl_service.service.components
 
 import android.content.Intent
-import io.matthewnelson.topl_service.service.ActionCommands.ServiceActionObject
-import io.matthewnelson.topl_service.service.ActionCommands.ServiceActionObjectGetter
+import io.matthewnelson.topl_service.service.BaseService
+import io.matthewnelson.topl_service.service.components.ActionCommands.ServiceActionObject
+import io.matthewnelson.topl_service.service.components.ActionCommands.ServiceActionObjectGetter
 import io.matthewnelson.topl_service.util.ServiceConsts
 import kotlinx.coroutines.*
